[DROP] Add my elisp comments in `pr.cc`

+++ src/benchmark.h
@@ -28,7 +28,7 @@ Various helper functions to ease writing of kernels
 
 
 // Default type signatures for commonly used types
-typedef int32_t NodeID;
+typedef uint64_t NodeID;
 typedef int32_t WeightT;
 typedef NodeWeight<NodeID, WeightT> WNode;
